Update QL720nw driver to QL820nwb driver
The QL820nwb label printer is the direct replacement for the QL720nw printer which has now been discontinued. The main difference being that the QL820nwb does not have a DB9 serial port to support direct UART connection to the imp005, but it does have USB. This USB driver for the QL720nw works with the QL720nw UART Driver, but to be compatible with the QL820nwb the PID needed to be changed to match the QL820nwb (0x209D). Comments have also been updated to say QL820nwb instead of QL720nw.